a little more

+++ code/modules/mob/living/brain/brain_item.dm
@@ -247,6 +247,8 @@
 	return ..()
 
 /obj/item/organ/brain/mmi_holder/Insert(mob/living/carbon/C, special = 0, no_id_transfer = FALSE)
+	if(special)
+		return
 	if(!stored_mmi)
 		qdel(src)
 		return
